+++ ChangeLog.md
@@ -1,6 +1,12 @@
 XP static initializer blocks - ChangeLog
 ========================================
 
+## ?.?.? / ????-??-??
+
+* Allowed multiple static initializer blocks, just like NodeJS does, see
+  https://v8.dev/features/class-static-initializer-blocks#multiple-blocks
+  (@thekid)
+
 ## 0.1.0 / 2021-12-19
 
 * Hello World! First release - @thekid